Extend updateBindings to set 'x-db-map' attribute
Extended the 'updateBindings' utility to not only set 'x-type' but also 'x-db-map' based on mapped properties. This is useful for applications where 'x-db-map' attribute is required to correspond with database mappings. The 'selectorsWithType' parameter is expanded to include 'dbMap' value, making it easier to update all bindings in one operation. This refactor enhances functionality and efficiency.