Refactor fetch/POST request with additional data & error logging
In the UI editor's loader route, the fetch request has been modified to a POST method with additional client-related data being sent with it for more specificity in the response. Also, the 'selectorsWithType' import was replaced with a second fetch request from an external API, making the process dynamic and more extensible than the static import. Lastly, error messages sent back to the client are now appended with the actual error object, providing more detail on what went wrong in case of an error.