Added targetFile parameter to retrieval functions
This commit introduces a 'targetFile' parameter to functions involved in fetching and saving data, allowing for more flexible file targeting instead of hardcoding 'index.html'. It also includes some code style changes such as semicolon addition for consistency. This change will improve the flexibility of the code and make it more reusable and testable.