Update README with instructions to move HTML templates
This commit adds an important detail to the README instructions. It instructs the user to move the **index.html**, **sms.html** and **email.html** from the **template folder** to the **"template" s3 bucket** which they have to create. This step is crucial to the proper setup and functioning of the local environment.